Add files via upload
This is all the code I've been working with. Some of it was written by me, but other things like colex, kthperm and the functions that generate random numbers were not. I am working on transcribing this code into python. The basic method for running simulations is to use generateSample to get a sample of elections, and then pass the sample into one of the tally methods. Eventually I'll write methods that specifically take in the outcomes of simulations and check how many times the given outcomes violate particular fairness criteria, as up to this point I have been doing that through the command line without built in functions.